developping LLM quiz generator (playground)/code parts for later usage or evaluation/splitted_code/requirements.txt
python-dotenv>=1.0.0
openai>=1.3.0
instructor>=0.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.2 